+++ History.md
@@ -1,3 +1,5 @@
+* add the option to preserve variables in the output
+
 3.0.0 / 2014-04-17
 ==================